ioboard.h: Add alias for IO Board signal of SC-SAT1 Main OBC.
The user IO assignments are as follows.

UIO3_00: UART_DE
This port is assigned to "Driver enable (DE)" on the RS-485 Transceiver.

UIO3_01: UART_RE_B
This port is assigned to "Receiver enable (REB)" on the RS-485
Transceiver. This signal is active low.

UIO3_O2: 3V3_SYS_EN
This port is assigned to the redundant signal of VDD_3V3_SYS.

Signed-off-by: Daisuke Sasaki <daisuke@spacecubics.com>